warning C4100:  unreferenced formal parameter
"mouse location in window relative coordinates when the key was pressed"
http://techpubs.sgi.com/library/tpl/cgi-bin/getdoc.cgi?coll=linux&db=man&fname=/usr/share/catman/man3/glutKeyboardFunc.3xglut.html

I currently can think of no beneficial use for this data.  Where the user's (x, y) mouse position was on-screen when a key was pressed doesn't seem useful to me for any feature currently, although potentially in a future.