Do not use low-level WINAPI calls for printing strings on Windows.
The `print_literal` macro function I wrote was to use kernel32:WriteFile instead of MSVCRT to print strings, but that was back when I had plans to remove MSVCRT as a dependency altogether from the Windows build.  In most (at least, normal and practical) cases however, the C standard library should be respected, and I think there is no more need for this remnant of the past.  At this point, it seems wiser to embrace MSVCRT calls than it is to remove the dependency altogether for this sort of utility.