replaced Windows build files with XP64-compatible DDK script
It is more instantaneous to build outside of Microsoft Windows, but in case of testing on Windows, I think just this WDK batch file (which will link against MSVCRT.DLL) should be more than sufficient.  Besides, the Unix script `make.sh` can be emulated through Git Bash on Windows to do a MinGW build that way.